# app/Http/Controllers/Schemas/TallaInputSchema.php
<?php

namespace App\Http\Controllers\Schemas;

/**
 * @OA\Schema(
 *     schema="TallaInput",
 *     required={"nombre"},
 *     @OA\Property(property="nombre", type="string", example="XXL")
 * )
 */
class TallaInputSchema {}