# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoCamisetas API",
        "description": "API REST para gestión de inventario de camisetas de fútbol y clientes B2B. Proveedor mayorista de camisetas nacionales e internacionales.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Servidor de desarrollo local"
        }
    ],
    "paths": {
...
        "/tallas": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Listar todas las tallas",
                "operationId": "3eaa361866e7e29246bd0d7e09fd0d75",
                "responses": {
                    "200": {
                        "description": "Listado de tallas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Talla"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Crear una talla",
                "operationId": "084652df5aaae551b2e445cc923440c1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TallaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Talla creada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Talla"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/tallas/{id}": {
            "get": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Obtener una talla con sus camisetas",
                "operationId": "e740ab9d2686678eeba210af9e46700d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de talla",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Talla"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Actualizar una talla",
                "operationId": "ead2c6582ff284b2c3c8fc2cc3903e5d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TallaInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Talla actualizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Talla"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tallas"
                ],
                "summary": "Eliminar una talla",
                "operationId": "89c6e70e8ac16e66d9ae93179e36e7cf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Talla eliminada"
                    },
                    "404": {
                        "description": "Talla no encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "TallaInput": {
                "required": [
                    "nombre"
                ],
                "properties": {
                    "nombre": {
                        "type": "string",
                        "example": "XXL"
                    }
                },
                "type": "object"
            },
            "Talla": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "M"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}